[modify] factory를 fixture로 추가
session 을 통한 데이터 생성
