--- src/tests/test_related_model.py
@@ -1,8 +1,0 @@
-from src.model.models import User, Todos
-
-
-def test_related_model_should_connect():
-    us = User()
-    td = Todos()
-    us.todos.append(td)
-    assert us is not None
